+++ fsutil/fs.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-// Package fsutil provides methods for working with files on posix compatible systems
+// Package fsutil provides methods for working with files on POSIX compatible systems
 package fsutil
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -24,23 +24,28 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 const (
-	_IFMT   = 0170000
-	_IFSOCK = 0140000
-	_IFLNK  = 0120000
-	_IFREG  = 0100000
-	_IFBLK  = 0060000
-	_IFDIR  = 0040000
-	_IFCHR  = 0020000
-	_IRUSR  = 00400
-	_IWUSR  = 00200
-	_IXUSR  = 00100
-	_IRGRP  = 00040
-	_IWGRP  = 00020
-	_IXGRP  = 00010
-	_IROTH  = 00004
-	_IWOTH  = 00002
-	_IXOTH  = 00001
+	_IFMT   = 0xf000
+	_IFSOCK = 0xc000
+	_IFLNK  = 0xa000
+	_IFREG  = 0x8000
+	_IFBLK  = 0x6000
+	_IFDIR  = 0x4000
+	_IFCHR  = 0x2000
+	_IRUSR  = 0x100
+	_IWUSR  = 0x80
+	_IXUSR  = 0x40
+	_IRGRP  = 0x20
+	_IWGRP  = 0x10
+	_IXGRP  = 0x8
+	_IROTH  = 0x4
+	_IWOTH  = 0x2
+	_IXOTH  = 0x1
 )
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// ErrEmptyPath error
+var ErrEmptyPath = errors.New("Path is empty")
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -86,7 +91,7 @@ func CheckPerms(props, path string) bool {
 			}
 
 		case 'L':
-			if stat.Mode&_IFMT != _IFLNK {
+			if !IsLink(path) {
 				return false
 			}
 
@@ -165,6 +170,10 @@ func IsExist(path string) bool {
 
 // IsRegular check if target is regular file or not
 func IsRegular(path string) bool {
+	if path == "" {
+		return false
+	}
+
 	path = PATH.Clean(path)
 	mode := getMode(path)
 
@@ -177,6 +186,10 @@ func IsRegular(path string) bool {
 
 // IsSocket check if target is socket or not
 func IsSocket(path string) bool {
+	if path == "" {
+		return false
+	}
+
 	path = PATH.Clean(path)
 	mode := getMode(path)
 
@@ -189,6 +202,10 @@ func IsSocket(path string) bool {
 
 // IsBlockDevice check if target is block device or not
 func IsBlockDevice(path string) bool {
+	if path == "" {
+		return false
+	}
+
 	path = PATH.Clean(path)
 	mode := getMode(path)
 
@@ -201,6 +218,10 @@ func IsBlockDevice(path string) bool {
 
 // IsCharacterDevice check if target is character device or not
 func IsCharacterDevice(path string) bool {
+	if path == "" {
+		return false
+	}
+
 	path = PATH.Clean(path)
 	mode := getMode(path)
 
@@ -213,6 +234,10 @@ func IsCharacterDevice(path string) bool {
 
 // IsDir check if target is directory or not
 func IsDir(path string) bool {
+	if path == "" {
+		return false
+	}
+
 	path = PATH.Clean(path)
 	mode := getMode(path)
 
@@ -225,14 +250,16 @@ func IsDir(path string) bool {
 
 // IsLink check if file is link or not
 func IsLink(path string) bool {
-	path = PATH.Clean(path)
-	mode := getMode(path)
-
-	if mode == 0 {
+	if path == "" {
 		return false
 	}
 
-	return mode&_IFMT == _IFLNK
+	path = PATH.Clean(path)
+
+	var buf = make([]byte, 1)
+	_, err := syscall.Readlink(path, buf)
+
+	return err == nil
 }
 
 // IsReadable check if file is readable or not
@@ -318,7 +345,7 @@ func IsNonEmpty(path string) bool {
 
 	path = PATH.Clean(path)
 
-	return GetSize(path) != 0
+	return GetSize(path) > 0
 }
 
 // IsEmptyDir check if directory empty or not
@@ -349,7 +376,7 @@ func IsEmptyDir(path string) bool {
 // GetOwner return object owner pid and gid
 func GetOwner(path string) (int, int, error) {
 	if path == "" {
-		return -1, -1, errors.New("Path is empty")
+		return -1, -1, ErrEmptyPath
 	}
 
 	path = PATH.Clean(path)
@@ -367,6 +394,10 @@ func GetOwner(path string) (int, int, error) {
 
 // GetATime return time of last access
 func GetATime(path string) (time.Time, error) {
+	if path == "" {
+		return time.Time{}, ErrEmptyPath
+	}
+
 	path = PATH.Clean(path)
 
 	atime, _, _, err := GetTimes(path)
@@ -376,6 +407,10 @@ func GetATime(path string) (time.Time, error) {
 
 // GetCTime return time of creation
 func GetCTime(path string) (time.Time, error) {
+	if path == "" {
+		return time.Time{}, ErrEmptyPath
+	}
+
 	path = PATH.Clean(path)
 
 	_, _, ctime, err := GetTimes(path)
@@ -385,6 +420,10 @@ func GetCTime(path string) (time.Time, error) {
 
 // GetMTime return time of modification
 func GetMTime(path string) (time.Time, error) {
+	if path == "" {
+		return time.Time{}, ErrEmptyPath
+	}
+
 	path = PATH.Clean(path)
 
 	_, mtime, _, err := GetTimes(path)
@@ -395,7 +434,7 @@ func GetMTime(path string) (time.Time, error) {
 // GetSize return file size in bytes
 func GetSize(path string) int64 {
 	if path == "" {
-		return 0
+		return -1
 	}
 
 	path = PATH.Clean(path)
@@ -405,25 +444,26 @@ func GetSize(path string) int64 {
 	err := syscall.Stat(path, stat)
 
 	if err != nil {
-		return 0
+		return -1
 	}
 
 	return stat.Size
 }
 
-// GetPerm return file permissions
-func GetPerm(path string) os.FileMode {
+// GetPerms return file permissions
+func GetPerms(path string) os.FileMode {
+	if path == "" {
+		return 0
+	}
+
 	path = PATH.Clean(path)
+
 	return os.FileMode(getMode(path) & 0777)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func getMode(path string) uint32 {
-	if path == "" {
-		return 0
-	}
-
 	var stat = &syscall.Stat_t{}
 
 	err := syscall.Stat(path, stat)
@@ -440,10 +480,11 @@ func isReadableStat(stat *syscall.Stat_t, uid int, gids []int) bool {
 		return true
 	}
 
-	switch {
-	case stat.Mode&_IROTH == _IROTH:
+	if stat.Mode&_IROTH == _IROTH {
 		return true
-	case stat.Mode&_IRUSR == _IRUSR && uid == int(stat.Uid):
+	}
+
+	if stat.Mode&_IRUSR == _IRUSR && uid == int(stat.Uid) {
 		return true
 	}
 
@@ -461,10 +502,11 @@ func isWritableStat(stat *syscall.Stat_t, uid int, gids []int) bool {
 		return true
 	}
 
-	switch {
-	case stat.Mode&_IWOTH == _IWOTH:
+	if stat.Mode&_IWOTH == _IWOTH {
 		return true
-	case stat.Mode&_IWUSR == _IWUSR && uid == int(stat.Uid):
+	}
+
+	if stat.Mode&_IWUSR == _IWUSR && uid == int(stat.Uid) {
 		return true
 	}
 
@@ -482,10 +524,11 @@ func isExecutableStat(stat *syscall.Stat_t, uid int, gids []int) bool {
 		return true
 	}
 
-	switch {
-	case stat.Mode&_IXOTH == _IXOTH:
+	if stat.Mode&_IXOTH == _IXOTH {
 		return true
-	case stat.Mode&_IXUSR == _IXUSR && uid == int(stat.Uid):
+	}
+
+	if stat.Mode&_IXUSR == _IXUSR && uid == int(stat.Uid) {
 		return true
 	}
 
@@ -500,7 +543,7 @@ func isExecutableStat(stat *syscall.Stat_t, uid int, gids []int) bool {
 
 func getGIDList(user *system.User) []int {
 	if user == nil {
-		return []int{}
+		return nil
 	}
 
 	var result []int