+++ httputil/httputil.go
@@ -1,4 +1,4 @@
-// Package httputil provides methods for working with http request/responses
+// Package httputil provides methods for working with HTTP request/responses
 package httputil
 
 // ////////////////////////////////////////////////////////////////////////////////// //