+++ jsonutil/jsonutil_test.go
@@ -85,7 +85,7 @@ func (s *JSONSuite) TestEncoding(c *C) {
 	err := EncodeToFile(jsonFile, testStruct, 0640)
 
 	c.Assert(err, IsNil)
-	c.Assert(fsutil.GetPerm(jsonFile), Equals, os.FileMode(0640))
+	c.Assert(fsutil.GetPerms(jsonFile), Equals, os.FileMode(0640))
 
 	data, err := ioutil.ReadFile(jsonFile)
 
@@ -94,6 +94,10 @@ func (s *JSONSuite) TestEncoding(c *C) {
 	c.Assert(string(data), Equals, _JSON_DATA)
 
 	err = EncodeToFile("/test.json", testStruct)
+
+	c.Assert(err, NotNil)
+
+	err = EncodeToFile(jsonFile, map[float64]int{3.14: 123})
 
 	c.Assert(err, NotNil)
 }