+++ terminal/window/size_posix.go
@@ -1,6 +1,7 @@
 // +build !windows
 
-package terminal
+// Package window provides methods for working terminal window
+package window
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -26,25 +27,38 @@ type winsize struct {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+// tty is path to tty device file
+var tty = "/dev/tty"
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // GetSize return window width and height
 func GetSize() (int, int) {
-	var tty *os.File
-
-	tty, err := os.OpenFile("/dev/tty", syscall.O_RDONLY, 0)
+	t, err := os.OpenFile(tty, syscall.O_RDONLY, 0)
 
 	if err != nil {
 		return -1, -1
 	}
 
-	defer tty.Close()
-
 	var sz winsize
 
 	_, _, _ = syscall.Syscall(
-		syscall.SYS_IOCTL, tty.Fd(),
+		syscall.SYS_IOCTL, t.Fd(),
 		uintptr(syscall.TIOCGWINSZ),
 		uintptr(unsafe.Pointer(&sz)),
 	)
 
 	return int(sz.cols), int(sz.rows)
+}
+
+// GetWidth return window width
+func GetWidth() int {
+	w, _ := GetSize()
+	return w
+}
+
+// GetHeight return window height
+func GetHeight() int {
+	_, h := GetSize()
+	return h
 }