+++ terminal/window/size_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package terminal
+package window
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -12,4 +12,14 @@ package terminal
 // GetSize return window width and height
 func GetSize() (int, int) {
 	return -1, -1
+}
+
+// GetWidth return window width
+func GetWidth() int {
+	return -1
+}
+
+// GetHeight return window height
+func GetHeight() int {
+	return -1
 }